+++ src/main/java/fr/univartois/butinfo/couleurs/PaletteRose.java
@@ -6,8 +6,9 @@ public class PaletteRose implements IStrategieCouleurs {
 
 	@Override
 	public Color palette(int nbItMax, int nbIterations) {
-		int ratio = nbIterations/nbItMax;
+		float ratio =(float) nbIterations/nbItMax;
 		return new Color(ratio,ratio/2,ratio);
 	}
+
 
 }
